refactor: plugins_robot
