+++ packages/plugins/robot/index.ts
@@ -11,7 +11,7 @@
  */
 
 import entry from './src/Main.vue'
-import metaData from './meta.js'
+import metaData from './meta'
 import './src/styles/vars.less'
 
 export default {
